Moved PrinterWidgetViewModel to formentry package
